+++ src/app/(blog)/blog/[slug]/page.tsx
@@ -26,6 +26,9 @@ export async function generateMetadata(
   return {
     title: post.title,
     description: post.excerpt,
+    alternates: {
+      canonical: `/blog/${post.slug}`,
+    },
     openGraph: {
       title: `${post.title} | FeedSport Blog`,
       description: post.excerpt,
