+++ src/app/(main)/contact/page.tsx
@@ -1,185 +1,14 @@
-'use client';
+import ContactPageClient from './ContactPageClient';
+import type { Metadata } from 'next';
 
-import { useState } from 'react';
-import { motion } from 'framer-motion';
-import { FiSend, FiUser, FiMail, FiMessageSquare } from 'react-icons/fi';
-import SecondaryHero from '@/components/common/SecondaryHero';
-import { Metadata } from 'next';
-
-// This is a client component, so we can't export metadata directly.
-// We would typically set this in the layout or a parent server component.
-// For now, we'll keep the static metadata in the root layout.
+export const metadata: Metadata = {
+  title: 'Contact Us',
+  description: 'Get in touch with FeedSport for expert advice on livestock nutrition and feed formulations.',
+  alternates: {
+    canonical: '/contact',
+  },
+};
 
 export default function ContactPage() {
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    message: ''
-  });
-
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ): void => {
-    const { id, value } = e.target;
-    setFormData(prev => ({ ...prev, [id]: value }));
-  };
-
-  const handleSubmit = (e: React.FormEvent<HTMLFormElement>): void => {
-    e.preventDefault();
-    // Handle form submission
-    console.log('Form submitted:', formData);
-  };
-
-  return (
-    <>
-      <SecondaryHero
-        title="Let's Connect"
-        subtitle="Our team is ready to help with your livestock nutrition needs. Reach out and we'll respond within 24 hours."
-      />
-      <main className="bg-gradient-to-b from-green-50 to-white py-16 px-4">
-        <div className="container mx-auto max-w-6xl">
-
-          <div className="grid grid-cols-1 lg:grid-cols-2 gap-12 items-center">
-            <motion.form
-              onSubmit={handleSubmit}
-              initial={{ opacity: 0, x: -20 }}
-              animate={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.5, delay: 0.2 }}
-              className="bg-white rounded-2xl shadow-xl p-8 lg:p-10"
-            >
-              <div className="space-y-6">
-                <div className="space-y-1">
-                  <label htmlFor="name" className="text-sm font-medium text-gray-700 flex items-center">
-                    <FiUser className="mr-2 text-green-600" />
-                    Your Name
-                  </label>
-                  <input
-                    type="text"
-                    id="name"
-                    autoComplete="name"
-                    value={formData.name}
-                    onChange={handleChange}
-                    className="w-full px-4 py-3 border border-gray-200 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent transition-all"
-                    placeholder="John Dube"
-                    required
-                  />
-                </div>
-
-                <div className="space-y-1">
-                  <label htmlFor="email" className="text-sm font-medium text-gray-700 flex items-center">
-                    <FiMail className="mr-2 text-green-600" />
-                    Email Address
-                  </label>
-                  <input
-                    type="email"
-                    id="email"
-                    autoComplete="email"
-                    value={formData.email}
-                    onChange={handleChange}
-                    className="w-full px-4 py-3 border border-gray-200 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent transition-all"
-                    placeholder="john@farm.co.zw"
-                    required
-                  />
-                </div>
-
-                <div className="space-y-1">
-                  <label htmlFor="message" className="text-sm font-medium text-gray-700 flex items-center">
-                    <FiMessageSquare className="mr-2 text-green-600" />
-                    Your Message
-                  </label>
-                  <textarea
-                    id="message"
-                    value={formData.message}
-                    onChange={handleChange}
-                    className="w-full px-4 py-3 border border-gray-200 rounded-lg focus:ring-2 focus:ring-green-500 focus:border-transparent transition-all min-h-[150px]"
-                    placeholder="How can we help you today?"
-                    required
-                  />
-                </div>
-
-                <motion.button
-                  whileHover={{ scale: 1.02 }}
-                  whileTap={{ scale: 0.98 }}
-                  type="submit"
-                  className="w-full bg-gradient-to-r from-green-600 to-green-700 text-white font-medium py-3 px-6 rounded-lg flex items-center justify-center space-x-2"
-                >
-                  <FiSend />
-                  <span>Send Message</span>
-                </motion.button>
-              </div>
-            </motion.form>
-
-            <motion.div
-              initial={{ opacity: 0, x: 20 }}
-              animate={{ opacity: 1, x: 0 }}
-              transition={{ duration: 0.5, delay: 0.4 }}
-              className="space-y-8"
-            >
-              <div className="bg-white/80 backdrop-blur-sm rounded-2xl shadow-lg p-8 border border-gray-100">
-                <h2 className="text-xl font-bold text-gray-800 mb-4">Other Ways to Reach Us</h2>
-
-                <div className="space-y-4">
-                  <div className="flex items-start">
-                    <div className="bg-green-100 p-3 rounded-full mr-4">
-                      <FiMail className="text-green-600 text-xl" />
-                    </div>
-                    <div>
-                      <h3 className="font-medium text-gray-700">Email</h3>
-                      <a href="mailto:support@feedsport.co.zw" className="text-green-600 hover:underline">
-                        support@feedsport.co.zw
-                      </a>
-                    </div>
-                  </div>
-
-                  <div className="flex items-start">
-                    <div className="bg-green-100 p-3 rounded-full mr-4">
-                      <svg className="w-5 h-5 text-green-600" fill="currentColor" viewBox="0 0 24 24">
-                        <path d="M20 22.621l-3.521-6.795c-.008.004-1.974.97-2.064 1.011-2.24 1.086-6.799-7.82-4.609-8.994l2.083-1.026-3.493-6.817-2.106 1.039c-7.202 3.755 4.233 25.982 11.6 22.615.121-.055 2.102-1.029 2.11-1.033z" />
-                      </svg>
-                    </div>
-                    <div>
-                      <h3 className="font-medium text-gray-700">Phone</h3>
-                      <a href="tel:+263774684534" className="text-green-600 hover:underline">
-                        +263 77 468 4534
-                      </a>
-                    </div>
-                  </div>
-
-                  <div className="flex items-start">
-                    <div className="bg-green-100 p-3 rounded-full mr-4">
-                      <svg className="w-5 h-5 text-green-600" fill="currentColor" viewBox="0 0 24 24">
-                        <path d="M12 2C8.13 2 5 5.13 5 9c0 5.25 7 13 7 13s7-7.75 7-13c0-3.87-3.13-7-7-7zm0 9.5c-1.38 0-2.5-1.12-2.5-2.5s1.12-2.5 2.5-2.5 2.5 1.12 2.5 2.5-1.12 2.5-2.5 2.5z" />
-                      </svg>
-                    </div>
-                    <div>
-                      <h3 className="font-medium text-gray-700">Location</h3>
-                      <address className="not-italic text-gray-600">2 Off William Pollet Drive, Borrowdale, Harare, Zimbabwe</address>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div className="bg-white/80 backdrop-blur-sm rounded-2xl shadow-lg p-8 border border-gray-100">
-                <h2 className="text-xl font-bold text-gray-800 mb-4">Business Hours</h2>
-                <ul className="space-y-2">
-                  <li className="flex justify-between">
-                    <span className="text-gray-600">Monday - Friday</span>
-                    <span className="font-medium">8:00 AM - 5:00 PM</span>
-                  </li>
-                  <li className="flex justify-between">
-                    <span className="text-gray-600">Saturday</span>
-                    <span className="font-medium">9:00 AM - 1:00 PM</span>
-                  </li>
-                  <li className="flex justify-between">
-                    <span className="text-gray-600">Sunday</span>
-                    <span className="font-medium">Closed</span>
-                  </li>
-                </ul>
-              </div>
-            </motion.div>
-          </div>
-        </div>
-      </main>
-    </>
-  );
+  return <ContactPageClient />;
 }
