+++ src/app/(main)/formulations/FormulationsClient.tsx
@@ -1,0 +1,15 @@
+'use client';
+import AboutSection from "@/components/home/AboutSection";
+import FeaturesSection from "@/components/home/FeaturesSection";
+import FeedFormulationsStrip from "@/components/home/FormulationStripe";
+import HeroSection from "@/components/home/HeroSection";
+
+export default function FormulationsClient() {
+  return (
+    <main>
+      <HeroSection />
+      <FeedFormulationsStrip />
+      <AboutSection />
+    </main>
+  );
+}
