+++ src/app/(main)/formulations/page.tsx
@@ -1,15 +1,14 @@
-'use client';
-import AboutSection from "@/components/home/AboutSection";
-import FeaturesSection from "@/components/home/FeaturesSection";
-import FeedFormulationsStrip from "@/components/home/FormulationStripe";
-import HeroSection from "@/components/home/HeroSection";
+import FormulationsClient from './FormulationsClient';
+import type { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'Custom Feed Formulations',
+  description: 'Discover how FeedSport formulates precise diets for optimal livestock performance.',
+  alternates: {
+    canonical: '/formulations',
+  },
+};
 
 export default function FormulationsPage() {
-  return (
-    <main>
-      <HeroSection />
-      <FeedFormulationsStrip />
-      <AboutSection />
-    </main>
-  );
+  return <FormulationsClient />;
 }
